requirement.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
mysql-connector==2.2.9
odfpy==1.4.0
openpyxl==2.4.11
PyYAML==5.1
records==0.5.3
requests==2.21.0
SQLAlchemy==1.3.3
tablib==0.13.0
urllib3==1.24.3
xlrd==1.2.0
xlwt==1.3.0
